+++ Cargo.toml
@@ -6,4 +6,5 @@ members = [
 	"oxide-auth-iron",
 	"oxide-auth-rocket",
 	"oxide-auth-rouille",
+	"oxide-auth-ring",
 ]
