+++ oxide-auth-ring/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "oxide-auth-ring"
+version = "0.1.0"
+authors = ["asonix <asonix@asonix.dog>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base64 = "0.9"
+chrono = "0.4.2"
+oxide-auth = { version = "0.5.0-preview.0", path = "../oxide-auth" }
+serde = "1.0"
+serde_derive = "1.0"
+serde_json = "1.0"
+ring = ">=0.13,<0.15"
+rmp-serde = "^0.13"
+url = "1.7"
