+++ oxide-auth/Cargo.toml
@@ -20,7 +20,6 @@ chrono = "0.4.2"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-ring = ">=0.13,<0.15"
 rmp-serde = "^0.13"
 url = "1.7"
 
@@ -28,6 +27,7 @@ url = "1.7"
 
 [dev-dependencies]
 reqwest = "^0.9"
+oxide-auth-ring = { version = "0.1.0", path = "../oxide-auth-ring" }
 
 [package.metadata.docs.rs]
 features = []
