--- oxide-auth/src/code_grant/extensions/mod.rs
@@ -1,4 +1,0 @@
-//! Provides standard extensions to the OAuth process.
-mod pkce;
-
-pub use self::pkce::Pkce;
