+++ oxide-auth/src/primitives/generator.rs
@@ -13,15 +13,7 @@ use super::grant::{Value, Extensions, Grant};
 use super::{Url, Time};
 use super::scope::Scope;
 
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::sync::Arc;
-
-use base64::{encode, decode};
-use ring::digest::SHA256;
-use ring::rand::{SystemRandom, SecureRandom};
-use ring::hmac::SigningKey;
-use rmp_serde;
+use std:: collections::HashMap;
 
 /// Generic token for a specific grant.
 ///
@@ -44,58 +36,8 @@ pub trait TagGrant {
     fn tag(&mut self, usage: u64, grant: &Grant) -> Result<String, ()>;
 }
 
-/// Generates tokens from random bytes.
-///
-/// Each byte is chosen randomly from the basic `rand::thread_rng`. This generator will always
-/// succeed.
-pub struct RandomGenerator {
-    random: SystemRandom,
-    len: usize
-}
-
-impl RandomGenerator {
-    /// Generates tokens with a specific byte length.
-    pub fn new(length: usize) -> RandomGenerator {
-        RandomGenerator {
-            random: SystemRandom::new(),
-            len: length
-        }
-    }
-
-    fn generate(&self) -> String {
-        let mut result = vec![0; self.len];
-        self.random.fill(result.as_mut_slice())
-            .expect("Failed to generate random token");
-        encode(&result)
-    }
-}
-
-/// Generates tokens by signing its specifics with a private key.
-///
-/// Tokens produced by the generator include a serialized version of the grant followed by an HMAC
-/// signature.  Since data is not encrypted, this token generator will ERROR if any private
-/// extension is present in the grant.
-///
-/// The actual generator is given by a `TaggedAssertion` from `Assertion::tag` which enables
-/// signing the same grant for different uses, i.e. separating authorization from bearer grants and
-/// refresh tokens.
-pub struct Assertion {
-    secret: SigningKey,
-}
-
-/// The cryptographic suite ensuring integrity of tokens.
-pub enum AssertionKind {
-    /// Uses [HMAC (RFC 2104)][HMAC] with [SHA-256 (FIPS 180-4)][SHA256] hash.
-    ///
-    /// [HMAC]: https://tools.ietf.org/html/rfc2104
-    /// [SHA256]: https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.180-4.pdf
-    HmacSha256,
-    #[doc(hidden)]
-    __NonExhaustive,
-}
-
 #[derive(Serialize, Deserialize)]
-struct SerdeAssertionGrant {
+pub struct SerdeAssertionGrant {
     /// Identifies the owner of the resource.
     owner_id: String,
 
@@ -119,108 +61,6 @@ struct SerdeAssertionGrant {
     public_extensions: HashMap<String, Option<String>>,
 }
 
-#[derive(Serialize, Deserialize)]
-struct AssertGrant(Vec<u8>, Vec<u8>);
-
-/// Binds a tag to the data. The signature will be unique for data as well as the tag.
-pub struct TaggedAssertion<'a>(&'a Assertion, &'a str);
-
-impl Assertion {
-    /// Construct an assertion from a custom secret.
-    ///
-    /// If the key material mismatches the key length required by the selected hash algorithm then
-    /// padding or shortening of the supplied key material may be applied in the form dictated by
-    /// the signature type. See the respective standards.
-    ///
-    /// If future suites are added where this is not possible, his function may panic when supplied
-    /// with an incorrect key length.
-    pub fn new(kind: AssertionKind, key: &[u8]) -> Self {
-        let key = match kind {
-            AssertionKind::HmacSha256 => SigningKey::new(&SHA256, key),
-            AssertionKind::__NonExhaustive => unreachable!(),
-        };
-
-        Assertion {
-            secret: key,
-        }
-    }
-
-    /// Construct an assertion instance whose tokens are only valid for the program execution.
-    #[deprecated = "Use the correctly named `ephemeral` instead."]
-    #[doc(hidden)]
-    pub fn ephermal() -> Assertion {
-        Self::ephemeral()
-    }
-
-    /// Construct an assertion instance whose tokens are only valid for the program execution.
-    pub fn ephemeral() -> Self {
-        Assertion {
-            secret: SigningKey::generate(&SHA256, &SystemRandom::new()).unwrap(),
-        }
-    }
-
-    /// Get a reference to generator for the given tag.
-    pub fn tag<'a>(&'a self, tag: &'a str) -> TaggedAssertion<'a> {
-        TaggedAssertion(self, tag)
-    }
-
-    fn extract<'a>(&self, token: &'a str) -> Result<(Grant, String), ()> {
-        let decoded = decode(token).map_err(|_| ())?;
-        let assertion: AssertGrant = rmp_serde::from_slice(&decoded).map_err(|_| ())?;
-        ring::hmac::verify_with_own_key(&self.secret, &assertion.0, &assertion.1).map_err(|_| ())?;
-        let (_, serde_grant, tag): (u64, SerdeAssertionGrant, String)
-            = rmp_serde::from_slice(&assertion.0).map_err(|_| ())?;
-        Ok((serde_grant.grant(), tag))
-    }
-
-    fn signature(&self, data: &[u8]) -> ring::hmac::Signature {
-        ring::hmac::sign(&self.secret, data)
-    }
-
-    fn counted_signature(&self, counter: u64, grant: &Grant) 
-        -> Result<String, ()>
-    {
-        let serde_grant = SerdeAssertionGrant::try_from(grant)?;
-        let tosign = rmp_serde::to_vec(&(serde_grant, counter)).unwrap();
-        let signature = self.signature(&tosign);
-        Ok(base64::encode(&signature))
-    }
-
-    fn generate_tagged(&self, counter: u64, grant: &Grant, tag: &str) -> Result<String, ()> {
-        let serde_grant = SerdeAssertionGrant::try_from(grant)?;
-        let tosign = rmp_serde::to_vec(&(counter, serde_grant, tag)).unwrap();
-        let signature = self.signature(&tosign);
-        Ok(encode(&rmp_serde::to_vec(&AssertGrant(tosign, signature.as_ref().to_vec())).unwrap()))
-    }
-}
-
-impl<'a> TaggedAssertion<'a> {
-    /// Sign the grant for this usage.
-    ///
-    /// This commits to a token that can be used–according to the usage tag–while the endpoint can
-    /// trust in it belonging to the encoded grant. `counter` must be unique for each call to this
-    /// function, similar to an IV to prevent accidentally producing the same token for the same
-    /// grant (which may have multiple tokens). Note that the `tag` will be recovered and checked
-    /// while the IV will not.
-    pub fn sign(&self, counter: u64, grant: &Grant) -> Result<String, ()> {
-        self.0.generate_tagged(counter, grant, self.1)
-    }
-
-    /// Inverse operation of generate, retrieve the underlying token.
-    ///
-    /// Result in an Err if either the signature is invalid or if the tag does not match the
-    /// expected usage tag given to this assertion.
-    pub fn extract<'b>(&self, token: &'b str) -> Result<Grant, ()> {
-        self.0.extract(token).and_then(|(token, tag)| {
-            if tag == self.1 {
-                Ok(token)
-            } else {
-                Err(())
-            }
-        })
-    }
-}
-
 impl<'a, T: TagGrant + ?Sized + 'a> TagGrant for Box<T> {
     fn tag(&mut self, counter: u64, grant: &Grant) -> Result<String, ()> {
         (&mut **self).tag(counter, grant)
@@ -232,56 +72,6 @@ impl<'a, T: TagGrant + ?Sized + 'a> TagGrant for &'a mut T {
         (&mut **self).tag(counter, grant)
     }
 }
-
-impl TagGrant for RandomGenerator {
-    fn tag(&mut self, _: u64, _: &Grant) -> Result<String, ()> {
-        Ok(self.generate())
-    }
-}
-
-impl<'a> TagGrant for &'a RandomGenerator {
-    fn tag(&mut self, _: u64, _: &Grant) -> Result<String, ()> {
-        Ok(self.generate())
-    }
-}
-
-impl TagGrant for Rc<RandomGenerator> {
-    fn tag(&mut self, _: u64, _: &Grant) -> Result<String, ()> {
-        Ok(self.generate())
-    }
-}
-
-impl TagGrant for Arc<RandomGenerator> {
-    fn tag(&mut self, _: u64, _: &Grant) -> Result<String, ()> {
-        Ok(self.generate())
-    }
-}
-
-
-impl TagGrant for Assertion {
-    fn tag(&mut self, counter: u64, grant: &Grant) -> Result<String, ()> {
-        self.counted_signature(counter, grant)
-    }
-}
-
-impl<'a> TagGrant for &'a Assertion {
-    fn tag(&mut self, counter: u64, grant: &Grant) -> Result<String, ()> {
-        self.counted_signature(counter, grant)
-    }
-}
-
-impl TagGrant for Rc<Assertion> {
-    fn tag(&mut self, counter: u64, grant: &Grant) -> Result<String, ()> {
-        self.counted_signature(counter, grant)
-    }
-}
-
-impl TagGrant for Arc<Assertion> {
-    fn tag(&mut self, counter: u64, grant: &Grant) -> Result<String, ()> {
-        self.counted_signature(counter, grant)
-    }
-}
-
 
 mod scope_serde {
     use primitives::scope::Scope;
@@ -333,7 +123,7 @@ mod time_serde {
 }
 
 impl SerdeAssertionGrant {
-    fn try_from(grant: &Grant) -> Result<Self, ()> {
+    pub fn try_from(grant: &Grant) -> Result<Self, ()> {
         let mut public_extensions: HashMap<String, Option<String>> = HashMap::new();
 
         if grant.extensions.private().any(|_| true) {
@@ -354,7 +144,7 @@ impl SerdeAssertionGrant {
         })
     }
 
-    fn grant(self) -> Grant {
+    pub fn grant(self) -> Grant {
         let mut extensions = Extensions::new();
         for (name, content) in self.public_extensions.into_iter() {
             extensions.set_raw(name, Value::public(content))
@@ -372,7 +162,7 @@ impl SerdeAssertionGrant {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use oxide_auth_ring::generator::{Assertion, AssertionKind, RandomGenerator};
 
     #[test]
     #[allow(dead_code, unused)]
