+++ oxide-auth/tests/access_token.rs
@@ -1,18 +1,27 @@
-use primitives::authorizer::{AuthMap, Authorizer};
-use primitives::issuer::TokenMap;
-use primitives::grant::{Grant, Extensions};
-use primitives::registrar::{Client, ClientMap};
+extern crate oxide_auth;
+extern crate oxide_auth_ring;
+extern crate base64;
+extern crate chrono;
+extern crate serde_json;
+extern crate url;
 
-use frontends::simple::endpoint::access_token_flow;
+use oxide_auth::primitives::authorizer::{AuthMap, Authorizer};
+use oxide_auth::primitives::issuer::TokenMap;
+use oxide_auth::primitives::grant::{Grant, Extensions};
+use oxide_auth::primitives::registrar::{Client, ClientMap};
+use oxide_auth::frontends::simple::endpoint::access_token_flow;
+use oxide_auth_ring::registrar::Pbkdf2;
 
 use std::collections::HashMap;
 
-use base64;
 use chrono::{Utc, Duration};
-use serde_json;
 
-use super::{Body, CraftedRequest, CraftedResponse, Status, TestGenerator, ToSingleValueQuery};
-use super::defaults::*;
+#[path = "./mod.rs"]
+#[allow(dead_code)]
+mod helpers;
+
+use helpers::{Body, CraftedRequest, CraftedResponse, Status, TestGenerator, ToSingleValueQuery};
+use helpers::defaults::*;
 
 
 struct AccessTokenSetup {
@@ -26,6 +35,7 @@ struct AccessTokenSetup {
 impl AccessTokenSetup {
     fn private_client() -> Self {
         let mut registrar = ClientMap::new();
+        registrar.set_password_policy(Pbkdf2::default());
         let mut authorizer = AuthMap::new(TestGenerator("AuthToken".to_string()));
         let issuer = TokenMap::new(TestGenerator("AccessToken".to_string()));
 
@@ -44,7 +54,7 @@ impl AccessTokenSetup {
         };
 
         let authtoken = authorizer.authorize(authrequest).unwrap();
-        registrar.register_client(client);
+        registrar.register_client(client).unwrap();
 
         let basic_authorization = base64::encode(&format!("{}:{}",
             EXAMPLE_CLIENT_ID, EXAMPLE_PASSPHRASE));
@@ -60,6 +70,7 @@ impl AccessTokenSetup {
 
     fn public_client() -> Self {
         let mut registrar = ClientMap::new();
+        registrar.set_password_policy(Pbkdf2::default());
         let mut authorizer = AuthMap::new(TestGenerator("AuthToken".to_string()));
         let issuer = TokenMap::new(TestGenerator("AccessToken".to_string()));
 
@@ -77,7 +88,7 @@ impl AccessTokenSetup {
         };
 
         let authtoken = authorizer.authorize(authrequest).unwrap();
-        registrar.register_client(client);
+        registrar.register_client(client).unwrap();
 
         let basic_authorization = base64::encode(&format!("{}:{}",
             EXAMPLE_CLIENT_ID, EXAMPLE_PASSPHRASE));
@@ -180,7 +191,7 @@ fn access_valid_private() {
 // apply and would be counter intuitive as such information is not preserved in `url`.
 #[test]
 fn access_equivalent_url() {
-    use primitives::authorizer::Authorizer;
+    use oxide_auth::primitives::authorizer::Authorizer;
 
     const CLIENT_ID: &str = "ConfusingClient";
     const REDIRECT_URL: &str = "https://client.example";
@@ -192,7 +203,7 @@ fn access_equivalent_url() {
         REDIRECT_URL.parse().unwrap(),
         EXAMPLE_SCOPE.parse().unwrap());
 
-    setup.registrar.register_client(confusing_client);
+    setup.registrar.register_client(confusing_client).unwrap();
 
     let authrequest = Grant {
         client_id: CLIENT_ID.to_string(),
